I am going to fix the ThemeProvider path and consolidate some components.
