+++ src/app/components/FeaturesSection.tsx
