+++ src/app/components/HeroSection.tsx
