+++ src/app/components/LiveDiffDemo.tsx
